Add new option for max/min lost item pricing
In addition to a 'default' price when price is missing or zero,
we can also accommodate a range of prices by saying it should be
at least 'X' and not more than 'Y'.  This also allows you to
effectively set a fixed price for all lost items by setting min
and max to an equal amount.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Jeff Godin <jgodin@tadl.org>