Serials special status support for Items tab
The Serial Control interface prevents direct editing of item status, as
some status changes require additional updates and/or can create data
conflicts.

In the current state, though, we have no way to set three important
statuses: claimed, not held, and not published.  At some point, claims in
particular should be more than a simple status, but in the meantime, the
ability to set these should not be overlooked.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>