LP#1088561: remove current password (renew password issue with TPAC)
If a user chooses to have their browser remember their password for the
catalog, it can fill in the password to the change password form page
as hidden dots. This behavior was observed specifically with Firefox.

Instead, add autocomplete="off" attribute to the form to force users to
enter existing password to change to a new password as intended.

Signed-off-by: Simon Mai <hieu.mai@mnsu.edu>
Signed-off-by: Ben Shum <bshum@biblio.org>