LP 1088561 - part 2, fix all update forms to not show saved passwords
Bug re-opened due to locating a few other locations in my account area
where the password was being saved by browsers like Firefox instead of
requiring users to key in current password to make changes.

This patch covers those two additional cases and changes the quoting
around the first case to be more consistent with the rest of the line.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Simon Mai <hieu.mai@mnsu.edu>