Facets should not be sticky when users click "Search"
We were using a hidden form variable to track which facets were in use,
at the same time as relying on the GET params to actually display the
facets, and munging the links on the remaining facets to properly
add/subtract facets.

The result of the hidden form variable was that:

1) clicking the Search button would unexpectedly result in the facet
applying to the new search (wat)

2) applying two facets at once and then searching again would result
in the facet vars being treated as a raw array.

Thanks to Dan Pearl for inspiring this patch with his fix for the latter
problem, and Kathy Lussier's recognition of the basic problem.

Closes LP# 1092946 and 1091659.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>