Return calculated value for SIP recall_ok test
Determine the value for the "recall OK" testing for patron information
requests the same we we determine the value for ciruclations and holds.
Previously, the recall-OK test always returned false, presumably since
recalls are not currently supported in Evergreen SIP.

Note that like holds and circs, we determine whether a user has good
standing based on the users penalties and whether the user is active /
non-barred.  Though we could in some cases, we are not checking whether
the user has the required permission to perform specific tasks.

This feature is enabled via a new oils_sip.xml configuration option:

<option name='patron_calculate_recal_ok' value='true' />

This option is disabled by default.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>