LP #1092921: Display of/sorting by CN affixes in Simplified Hold Pull List
In the Simplified Hold Pull List interface, the "Call Number Label" column
displays the call number of the copy targeted by the hold in each row,
and it does so by combining affixes with the actual label field.
Sorting of this column actually uses the call number sortkey generated
in-database for this purpose.

Sometimes this means that sorting order doesn't match what you'd
expect to see with this column.  Opinions differ on desired behavior, so
now you can explicity add columns to the table for call number prefix
and suffix, and sort on either of those (potentially in combination with
other columns).

Amended from an earlier version of this commit that still didn't sort
CN prefix and suffix columns quite right.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>

Conflicts:
	Open-ILS/src/templates/circ/hold_pull_list.tt2

Signed-off-by: Ben Shum <bshum@biblio.org>