LP #1075167: Serials: Avoid infinite loop in holdings summarization
When receiving items with unworkable combinations of holdings and
patterns, the holdings summarization code can get stuck in an infinite
loop.  This solution prevents that, sparing the open-ils.serial
process and server resources, although it doesn't help the user out.

It will take further examination of the problem and reëvalution of how
we deal with problems reported by OpenILS::Utils::MFHD to help the
user avoid or resolve these situations.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>