ACQ API for setting the lineitem order identifier
API: open-ils.acq.lineitem.order_identifier.set

Given an existing lineitem_attr (typically a marc_attr), this will
create a matching local_attr to store the name and value and mark
the attr as the order_ident.  Any existing local_attr marked as
order_ident is removed.  When necessary, the value (isbn, issn, upc) is
added to the record first to create the marc attribute.  If the lineitem
is linked to a bib record and the user has the correct permissions, any
new data will also be applied to the bib record.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>