LP1054053 - Remove white padding around login inputs
Forget efforts to make the boxes bigger and smoother as it was in the past.
Instead, opting for a cleaner look by removing the extra white background
that was part of the padding surrounding the actual login input area. This
removes the feeling that the box area should have been bigger than it really
is.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Michael Peters <mrpeters@library.in.gov>