Z39.50 Batch Search/Overlay API
New API for performing Z39.50 queries for a set of bib records across a
set of Z39 sources.  Searches are performed in parallel (up to a
configurable maximum number of active searches and maximum results) and
results are stored in a vandelay MARC Import/Export queue.

Search fields and values are determined by Z39 index field maps.  At
search time, the caller provides a set of field maps to use for the
search. Search values for each bib record are found by following the
index maps to the indexed values in metabib.*_field_entry or
metabib.record_attr.

Signed-off-by: Bill Erickson <berick@esilibrary.com>