Default ACQ copies use workstation org as owning lib
When creating default copies for new lineitems or lineitems added to
POs, use the workstation org unit as the owning lib for each copy.  With
this, it's possible to order items for "here" without having to edit any
single copies.  This also resolves a confusing issue where opening the
copies in the copies grid made it appear that each copy had an
owning_lib applied (via the autofieldwidget) when none was yet set.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>