Acq: Line item batch update API method
This new method gives you a server-side way to make batch updates to
all the copies of a target set of lineitems.  Permissions are checked
for each target object.  The method returns OpenILS::Event instances
for all kinds of permission problems and other errors.  IDs of lineitems
affected are streamed back to the client in the successful case.

This also expands distribution formulae to contain more (optional)
fields so that they can be used to make updates to copies on all the
same fields that you can just specify individually.  IDL and interface
changes to support the new fields are included.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>