LP1054322 - libparent-perl not needed for Ubuntu Precise
The newer version of Perl 5.14 which comes with Ubuntu Precise already
contains the contents of libparent-perl, which is now an obsolete package.

Remove this dependency from the extra debs installed for Debian and Ubuntu
but then add to the extras for Debian Squeeze and create a new extras just
for Ubuntu Lucid so that it continues to be installed on those systems and
not on Ubuntu Precise.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>