Tweak Google Book Preview Display, Take 2
After some additional feedback, this commit makes three changes
to the Google Book preview link:

1) As a solution offered by Dan Scott, rather than move the
button, we simply scroll to the preview when the button is clicked.
This effectively makes it impossible to miss. (In the current
code it sometimes loads off the screen.)

2) Make the viewer display in the same style as other added
content.

3) Allow the added content header to function as an additional
toggle to turn the viewer on and off.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Ben Shum <bshum@biblio.org>