Don't populate the MFHD menu with distributions
If a distribution has no legacy record, it shouldn't get a menu
entry in the MFHD menus. A magic id of '-1' is used to indicate
this case, so we need to check for that when populating
mfhdDetails via the TPAC.

Also, entryNum must be an int, not a string, because it is
incremented prior to display. Removing the quotes solves this
display issue.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Ben Shum <bshum@biblio.org>