truncate fines to max fine amount (LP#1145284)
When the max fine amount is not a multiple of the recurring fine amount,
fines will actually max out at an amount greater than the max fine
value.  This commit adds YAOUS and some simple functionality to
optionally truncate the final amount billed to the max fine amount.

For example, if max fine is $5.00 and recurring fine is $0.30, the 17th
billing will bring the total amount billed to $5.10 (17 x $0.30),
thereby exceeding the max fine amount.  With this commit, if
circ.fines.truncate_to_max_fine is true, the final billing amount will
be reduced and the total amount billed will be $5.00.

Signed-off-by: Jeff Davis <jdavis@sitka.bclibraries.ca>
Signed-off-by: Mike Rylander <mrylander@gmail.com>