LP#1112723 Consider copy-level opac_visible flag when ranking
The current unAPI calls consider opac_visibility of the status,
but not the copy itself. If you have a record with copies set to
opac_visible = false, you might see few or no copies in the OPAC
results list, as these invisible copies may be the first five
returned, and they then crowd out the visible copies.

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Josh Stompro <stomproj@larl.org>
Signed-off-by: Ben Shum <bshum@biblio.org>