Acq: Line item batch update UI
This UI enhancement uses the API method in the previous commit to allow
users to apply changes to the copies on a selected set of line items.

When successful, the updater just reloads the whole page, but it can be
made smarter later, i.e. by refreshing only the affected line items rows
in the display.

The UI only appears on POs for now, and is partly disabled if the PO has
already been activated.  It could appear elsewhere if it seems sensible
to put it elsewhere.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>