QP search modifier '#deleted'
These tweaks involve an internal flag that's off by default. If you
want '#deleted', you'll have to turn it on explicitly at your site (and
reingest existing records) to get the functionality. If you don't need
'#deleted', you may prefer to leave the setting off so that your system
will be purged (as before) of useless (to you) metarecord mappings
taking up room in the database for deleted bibs.

[LFW] upgrade script; commit message; release note

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>