Move serials interfaces closer together
Back in (I believe) 2.2, we brought the dijit issuance editor into the
serial control interface with two goals: having less code in more places,
and easing the transition to an eventual unified interface which would
likely be more widget-based.

That experiment was well received, so this commit takes things a large
step further in that direction.  Basically, we are embedding the
alternate subscription editing interface into the serial control
scaffolding, which ultimately amounts to giving you some of the
benefits of both.  You get the more modern and maintainable code from
the alternate interfaces, but keep the context-awareness and familiar
navigation of the non-alternate serial control.

That said, this change is not meant to be the last word in serials,
but rather a way to at least come closer together while we still
continue to learn what structures provide the best experience.

Finally, some feature additions were necessary to the alternate serials
code in order to make this change as seamless as possible.  In some
cases, the new feature only works in the embedded version, but a new
'context' variable makes that simple to adjust as we move forward.
Changes include:
  - note editors on subscriptions and distributions
  - labeled dropdown for distribution summary options
  - and legacy record entry linkages

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>