Reify required authority fixed field names
Adam (and Art, by way of Adam's patch) added a goodly pile of fixed
field definitions that were lacking previously.  The change in this
commit simply changes the label of two (008/11 and 008/15) to those
in use by other parts of the code.  In particular, internally, we call
the field holding a thesaurus code value "Subj" and use that name during
ingest of authority records.  Therefore, that should not change.  This
required a subsequent change to the label of another, to SubjUse.

I've included a duplication in the backend-only (probably deprecated)
version of this mapping to allow the front-end labels to work.

Signed-off-by: Mike Rylander <mrylander@gmail.com>