Fix detection of empty Syndetics AC data elements
Due to a change in how the vendor responds to requests for data
elements which are not present for a requested title, we need to
change how we detect "empty" data elements in the "data_exists"
function.

First, we check for the new "No Data Available" title element, then
we check for the previously-used value of "error".

This change resolves the symptom where the catalog would display
tabs for "Author Notes", "Excerpt", "Table of Contents", on every
record in the catalog, even when clicking each tab would show that
the content was not present / available.

Signed-off-by: Jeff Godin <jgodin@tadl.org>
Signed-off-by: Ben Shum <bshum@biblio.org>