LP1103068 - Show titles and authors for pre-cat items
As noted in LP1103068, TPAC does not currently show anything when viewing a
pre-cataloged item in the my account summary of items checked out.

If the call number is -1 (pre-cat), then show the dummy title and dummy author
values, which fields are hopefully relevant as long as staff were good to put
something meaningful in those during creation of the pre-cat.

Also, wrap the title in a check for attrs.title.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Michael Peters <mpeters@emeralddata.net>