LP#1155313: Repair generation of label_sortkey for monograph_part entries
The evergreen.lpad_number_substrings function attempts to codify numeric
fields within labels.  It does this by finding the strings, padding them
to a given size, and replacing them in the source string.  For instance:
       3 => 0000000003
    15.4 => 00000000150000000004

The algorithm was fooled by repeated characters, like in 15.1:
    15.1 => 00000000000000000150000000001  INCORRECT

This change will result in the correct value.

Signed-off-by: Dan Pearl <dpearl@cwmars.org>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>