Undo holding lib to SRE linkage (alt dist editor)
Since the legacy record dropdown piggy-backs on the OPAC display, it
doesn't work in the alt interface (which has no clear OPAC context).

Worse, this leads to a Javascript error on the first editor load for
the distribution list, preventing subsequent editors.

A better fix will require more consideration, but for now, let's
revert to the unformatted display used in 2.3.

Also, this commit removes an unnecessary conditional (it was always
true).

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Mike Rylander <mrylander@gmail.com>