Use inline-block for opac-multiline-button style
Follow up for LP1102934 which introduced CSS for opac-multiline-button.

The button appeared noticeably larger if using the english text for
"FAQs" on the login help button. Using inline-block for the style shrinks
it to fit more appropriately based on actual amount of text instead of
leaving it an unusually large looking button.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Kyle Tomita <ktomita@catalystitservices.com>