Allow combined search to be optional per class
Relevance is thrown off for, in paticular, the keyword class when combined
search is used.  This is because the effect of an opaque blob of data, such
as the keyword|keyword index definition, is to applify the inclusion of spurious
(to the user) data in matching attempts.

This commit adds the ability to specify, per class, whether combined FTS should
be used, and turns this on for only the subject class by default.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Ben Shum <bshum@biblio.org>