LP 1169193 long-overdue automated processing
Middle layer components for marking circulations long overdue.  The
automation is controlled with an Action/Trigger event definition
(sample included in seed data), which uses a new MarkItemLongOverdue
reactor.  The reactor makes use of a new util function (AssetCommon.pm)
for doing the actual work and looking up the relevant org unit settings.

Inluded is a new PatronNotInCollections Validator, which can be used to
avoid marking circulations long overdue for patrons that are in
collections processing.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>