Correct subfield format for authority lookup
The Authority Control Set code expects the format of subfields
that it will be used to build a MARC.Field object to be of the
form: [[code,value],...]  This commit makes that true.

Additionally, dojo.filter over a list of lists can cause the
nested array set to be flattened.  So, instead, we loop directly.

Signed-off-by: Mike Rylander <mrylander@gmail.com>